# Conflicts:
#	Musicx.Presentation.Web/Musicx.Presentation.Web.Client/Components/GenreBadge.razor

+++ Musicx.Application.Web/Interfaces/UseCases/Specifics/IGetAlbumByGenreUseCase.cs
@@ -6,20 +6,20 @@ public interface IGetAlbumByGenreUseCase
 {
     Task<List<OutAlbum>> ExecuteAsync(long genreId, 
         int genreOptions,
-        int skip = 0,
-        int take = 100,
+        long skip = 0,
+        long take = 100,
         string query = "");
     Task<List<OutAlbum>> ExecuteAsync(OutGenre genre, int genreOptions,
-        int skip = 0,
-        int take = 100,
+        long skip = 0,
+        long take = 100,
         string query = "");
     
     List<OutAlbum> Execute(long genreId, int genreOptions,
-        int skip = 0,
-        int take = 100,
+        long skip = 0,
+        long take = 100,
         string query = "");
     List<OutAlbum> Execute(OutGenre genre, int genreOptions,
-        int skip = 0,
-        int take = 100,
+        long skip = 0,
+        long take = 100,
         string query = "");
 }
